+++ src/main/java/com/jug/GrowthLineFrame.java
@@ -5,7 +5,7 @@ package com.jug;
 
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.algorithm.componenttree.ComponentForest;
-import net.imglib2.type.numeric.real.DoubleType;
+import net.imglib2.type.numeric.real.FloatType;
 
 import com.jug.util.filteredcomponents.FilteredComponent;
 import com.jug.util.filteredcomponents.FilteredComponentTree;
@@ -19,14 +19,14 @@ import com.jug.util.filteredcomponents.FilteredComponentTree;
  *         series (2d+t) representation of an growth line is
  *         <code>GrowthLine</code>.
  */
-public class GrowthLineFrame extends AbstractGrowthLineFrame< FilteredComponent< DoubleType > > {
+public class GrowthLineFrame extends AbstractGrowthLineFrame< FilteredComponent< FloatType > > {
 
 	/**
 	 * @see com.jug.AbstractGrowthLineFrame#buildIntensityTree(net.imglib2.RandomAccessibleInterval)
 	 */
 	@Override
-	protected ComponentForest< FilteredComponent< DoubleType >> buildIntensityTree( final RandomAccessibleInterval< DoubleType > raiFkt ) {
-		return FilteredComponentTree.buildComponentTree( raiFkt, new DoubleType(), 3, Long.MAX_VALUE, true );
+	protected ComponentForest< FilteredComponent< FloatType >> buildIntensityTree( final RandomAccessibleInterval< FloatType > raiFkt ) {
+		return FilteredComponentTree.buildComponentTree( raiFkt, new FloatType(), 3, Long.MAX_VALUE, true );
 //		return MserTree.buildMserTree( raiFkt, MotherMachine.MIN_GAP_CONTRAST / 2.0, MotherMachine.MIN_CELL_LENGTH, Long.MAX_VALUE, 0.5, 0.33, true );
 	}
 
@@ -34,8 +34,8 @@ public class GrowthLineFrame extends AbstractGrowthLineFrame< FilteredComponent<
 	 * @see com.jug.AbstractGrowthLineFrame#buildParaMaxFlowSumTree(net.imglib2.RandomAccessibleInterval)
 	 */
 	@Override
-	protected ComponentForest< FilteredComponent< DoubleType >> buildParaMaxFlowSumTree( final RandomAccessibleInterval< DoubleType > raiFkt ) {
-		return FilteredComponentTree.buildComponentTree( raiFkt, new DoubleType(), 3, Long.MAX_VALUE, true );
+	protected ComponentForest< FilteredComponent< FloatType >> buildParaMaxFlowSumTree( final RandomAccessibleInterval< FloatType > raiFkt ) {
+		return FilteredComponentTree.buildComponentTree( raiFkt, new FloatType(), 3, Long.MAX_VALUE, true );
 	}
 
 }
